Add --ask option to prompt for further action
This option shall be used for several functions like upgrade, install,
remove etc. to prompt the user before potentially long lasting
operations or as small safety measure.

This commit only affects the upgrade command, but it is designed to be
usable by other commands.

Technically the --ask option only sets the environment variable
HOMEBREW_ASK to 1 which then can be checked against in each command. To
not affect further commands, the option is removed from the option list.

Credits to the Stack Overflow answer
http://stackoverflow.com/a/6278862/1532986
for the prompt() function.